feat: fix take case, adding batches and removing project name approach
